# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: sdds API
  version: 0.0.0
  description: HTTP JSON contract for the sdds API.
  license:
    name: UNLICENSED
    url: https://spdx.org/licenses/UNLICENSED.html
servers:
  - url: /
    description: Active API origin
security: []
...
paths:
...
  /v1/notes:
    get:
      operationId: listNotes
      summary: List recent notes
      tags:
        - Notes
      security:
        - bearerAuth: []
      parameters:
        - name: category_slug
          in: query
          required: false
          description: Optional active category slug used to narrow recent notes.
          schema:
            $ref: "#/components/schemas/CategorySlug"
      responses:
        "200":
          description: Recent notes.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListNotesResponse"
        "400":
          description: The request does not match the API contract.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: The API could not list notes.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: The request is missing a valid authenticated session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createNote
      summary: Create a note
      tags:
        - Notes
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNoteRequest"
      responses:
        "201":
          description: The note was created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "400":
          description: The request JSON or note content is invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: The request is missing a valid authenticated session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: The note request or image association conflicts with an existing request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: The request body is too large.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: The image association is temporarily unavailable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: The API could not create the note.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/notes/{note_id}:
    get:
      operationId: getNote
      summary: Get a note
      tags:
        - Notes
      security:
        - bearerAuth: []
      parameters:
        - name: note_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The note.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "400":
          description: The request does not match the API contract.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: The note was not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: The API could not get the note.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: The request is missing a valid authenticated session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /v1/notes/{note_id}/useful:
    put:
      operationId: markNoteUseful
      summary: Mark a note as useful
      tags:
        - Notes
      security:
        - bearerAuth: []
      parameters:
        - name: note_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: The note was marked useful.
        "401":
          description: The request is missing a valid authenticated session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: The note was not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: The API could not mark the note useful.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: unmarkNoteUseful
      summary: Remove a useful mark from a note
      tags:
        - Notes
      security:
        - bearerAuth: []
      parameters:
        - name: note_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: The useful mark was removed.
        "401":
          description: The request is missing a valid authenticated session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: The note was not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: The API could not remove the useful mark.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/search/notes:
    get:
      operationId: searchNotes
      summary: Search notes
      tags:
        - Notes
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: false
          description: Raw user search text. Missing, blank, or overlong values return invalid_search.
          schema:
            type: string
        - name: category_slug
          in: query
          required: false
          description: Optional active category slug used to narrow search results.
          schema:
            $ref: "#/components/schemas/CategorySlug"
      responses:
        "200":
          description: Matching notes.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchNotesResponse"
        "400":
          description: The search query is invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: The API could not search notes.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: The request is missing a valid authenticated session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    AuthorSummary:
      type: object
      additionalProperties: false
      required:
        - id
        - display_name
      properties:
        id:
          type: string
        display_name:
          type: string
...
    CategorySlug:
      type: string
...
    CreateNoteRequest:
      type: object
      additionalProperties: false
      required:
        - title
        - body
        - category_slug
        - client_request_id
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 120
        body:
          type: string
          minLength: 1
          maxLength: 4000
        category_slug:
          $ref: "#/components/schemas/CategorySlug"
        client_request_id:
          type: string
          minLength: 1
          maxLength: 128
        image_upload_ids:
          type: array
          uniqueItems: true
          maxItems: 1
          items:
            type: string
...
    ErrorCode:
      type: string
      x-enum-varnames:
        - ErrorCodeInternal
        - ErrorCodeForbidden
        - ErrorCodeInvalidAuth
        - ErrorCodeInvalidComment
        - ErrorCodeInvalidJSON
        - ErrorCodeInvalidNote
        - ErrorCodeInvalidReport
        - ErrorCodeInvalidSearch
        - ErrorCodeNotFound
        - ErrorCodeRateLimited
        - ErrorCodeRequestTooLarge
        - ErrorCodeUnauthenticated
        - ErrorCodeUsernameTaken
        - ErrorCodeInvalidMedia
        - ErrorCodeUnsupportedMediaType
        - ErrorCodeIdempotencyConflict
        - ErrorCodeUploadInProgress
        - ErrorCodeUploadExpired
        - ErrorCodeMediaStagingQuotaExceeded
        - ErrorCodeMediaStorageUnavailable
        - ErrorCodeMediaIntegrityError
        - ErrorCodeTooManyImages
        - ErrorCodeInvalidEvent
        - ErrorCodeInvalidEventBatch
      enum:
        - internal_error
        - forbidden
        - invalid_auth
        - invalid_comment
        - invalid_json
        - invalid_note
        - invalid_report
        - invalid_search
        - not_found
        - rate_limited
        - request_too_large
        - unauthenticated
        - username_taken
        - invalid_media
        - unsupported_media_type
        - idempotency_conflict
        - upload_in_progress
        - upload_expired
        - media_staging_quota_exceeded
        - media_storage_unavailable
        - media_integrity_error
        - too_many_images
        - invalid_event
        - invalid_event_batch
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - code
      properties:
        code:
          $ref: "#/components/schemas/ErrorCode"
        fields:
          type: array
          items:
            $ref: "#/components/schemas/ValidationProblem"
...
    SearchVersion:
      type: string
      enum: [fts5-v1]
    RetrievalSource:
      type: string
      enum: [lexical, semantic, hybrid]
...
    ListNotesResponse:
      type: object
      additionalProperties: false
      required:
        - notes
      properties:
        notes:
          type: array
          items:
            $ref: "#/components/schemas/Note"
    SearchNotesResponse:
      type: object
      additionalProperties: false
      required:
        - search_version
        - results
      properties:
        search_version:
          $ref: "#/components/schemas/SearchVersion"
        results:
          type: array
          items:
            $ref: "#/components/schemas/SearchNoteResult"
    SearchNoteResult:
      type: object
      additionalProperties: false
      required:
        - note
        - retrieval_source
      properties:
        note:
          $ref: "#/components/schemas/Note"
        retrieval_source:
          $ref: "#/components/schemas/RetrievalSource"
...
    NoteImage:
      type: object
      additionalProperties: false
      required:
        - id
        - url
        - content_type
        - byte_size
        - width
        - height
        - position
        - created_at
        - updated_at
      properties:
        id:
          type: string
        url:
          type: string
          minLength: 1
          format: uri-reference
        content_type:
          type: string
          enum:
            - image/jpeg
            - image/png
        byte_size:
          type: integer
          format: int64
          minimum: 1
        width:
          type: integer
          format: int32
          minimum: 1
        height:
          type: integer
          format: int32
          minimum: 1
        position:
          type: integer
          format: int32
          minimum: 0
        created_at:
          type: integer
          format: int64
        updated_at:
          type: integer
          format: int64
    Note:
      type: object
      additionalProperties: false
      required:
        - id
        - title
        - body
        - category_slug
        - author
        - created_at
        - updated_at
        - images
        - useful_count
        - useful_by_current_user
      properties:
        id:
          type: string
        title:
          type: string
        body:
          type: string
        category_slug:
          $ref: "#/components/schemas/CategorySlug"
        author:
          $ref: "#/components/schemas/AuthorSummary"
        created_at:
          type: integer
          format: int64
          description: Unix timestamp in milliseconds.
        updated_at:
          type: integer
          format: int64
          description: Unix timestamp in milliseconds.
        images:
          type: array
          items:
            $ref: "#/components/schemas/NoteImage"
        useful_count:
          type: integer
          format: int64
          minimum: 0
          description: Number of users who marked this note useful.
        useful_by_current_user:
          type: boolean
          description: Whether the current authenticated user marked this note useful.
    ValidationField:
      type: string
      x-enum-varnames:
        - ValidationFieldTitle
        - ValidationFieldBody
        - ValidationFieldCategorySlug
        - ValidationFieldQ
        - ValidationFieldUsername
        - ValidationFieldPassword
        - ValidationFieldDisplayName
        - ValidationFieldLimit
        - ValidationFieldCursor
        - ValidationFieldClientRequestID
        - ValidationFieldUploadRequestID
        - ValidationFieldImageUploadIDs
        - ValidationFieldFile
        - ValidationFieldTargetType
        - ValidationFieldTargetID
        - ValidationFieldReason
        - ValidationFieldDetails
      enum:
        - title
        - body
        - category_slug
        - q
        - username
        - password
        - display_name
        - limit
        - cursor
        - client_request_id
        - upload_request_id
        - image_upload_ids
        - file
        - target_type
        - target_id
        - reason
        - details
    ValidationProblem:
      type: object
      additionalProperties: false
      required:
        - field
        - code
      properties:
        field:
          $ref: "#/components/schemas/ValidationField"
        code:
          type: string
          x-enum-varnames:
            - ValidationProblemCodeRequired
            - ValidationProblemCodeTooShort
            - ValidationProblemCodeTooLong
            - ValidationProblemCodeUnknown
            - ValidationProblemCodeInvalid
            - ValidationProblemCodeTaken
          enum:
            - required
            - too_short
            - too_long
            - unknown
            - invalid
            - taken